refactors: address pr comments
